admin_panel/requirements.txt
django
django-async-orm
psycopg2-binary
openpyxl
aiogram>=3.0.0
python-dotenv
